Added first bunch of tests for package archive
