+++ testdata/foo
@@ -1,0 +1,1 @@
+bar
